Move new nodes to /standard
