+++ src/graph-nodes/standard/sinks/index.tsx
@@ -1,22 +1,5 @@
-import { any, number } from "superstruct";
+import { any } from "superstruct";
 import BaseNode from "../../../base-node";
-
-const Add = {
-  inputs: {
-    left: number(),
-    right: number(),
-  },
-  outputs: {
-    sum: ({ left, right }) => +left + +right,
-  },
-  Component: ({ data: { inputs, outputs } }) => {
-    return (
-      <BaseNode sources={inputs} sinks={outputs}>
-        <div>{"Add"}</div>
-      </BaseNode>
-    );
-  },
-};
 
 const Viewer = {
   inputs: {
@@ -35,6 +18,5 @@ const Viewer = {
 };
 
 export default {
-  Add,
   Viewer,
 };
